refactor: adds interfaces folder
