+++ src/commands/Hello.ts
@@ -5,7 +5,7 @@ import {
   InteractionResponse,
 } from 'discord.js';
 
-import { Command } from '../Command';
+import { Command } from '../interfaces/Command';
 
 export const Hello: Command = {
   name: 'hello',
