--- src/interfaces/Commands.ts
@@ -1,7 +1,0 @@
-import { FxTweet } from '../commands/FxTweet';
-import { Hello } from '../commands/Hello';
-
-export const Commands = [
-    Hello, 
-    FxTweet
-];
