+++ src/listeners/ready.ts
@@ -1,5 +1,5 @@
 import { Client, ActivityType } from 'discord.js';
-import { Commands } from '../interfaces/Commands';
+import { Commands } from '../Commands';
 
 export default (client: Client): void => {
   client.on('ready', async () => {
